Refactoring a few wrappers and adding exists() + keys() functions to helper
Previously, the entity module was responsible for most of the pre-wrapped calls for things like ``KEYS()`` or ``EXISTS()``. This change will move this functionality into the helper itself.

Based on the configuration of ``RedisentHelper.is_async``, these wrapper methods will correctly call the underlying ``*_async`` or ``*_sync`` implementation.

Added a bit of basic testing around this but more to come there.